modifying the tables with the extension prod
