+++ src/main/java/com/dataquadinc/controller/LogoutController.java
@@ -1,11 +1,8 @@
 package com.dataquadinc.controller;
 import com.dataquadinc.dto.LogoutResponseDTO;
-import com.dataquadinc.model.Roles;
 import com.dataquadinc.service.LogoutService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.Set;
 
 
 @CrossOrigin(origins = {"http://35.188.150.92", "http://192.168.0.140:3000", "http://192.168.0.139:3000"})
