+++ src/main/java/com/dataquadinc/dto/EmployeeWithRole.java
@@ -1,14 +1,10 @@
 package com.dataquadinc.dto;
 
-import com.dataquadinc.model.Roles;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.time.LocalDate;
-import java.util.List;
-import java.util.Set;
 
 @Data
 @NoArgsConstructor
