--- src/main/java/com/dataquadinc/model/UserType.java
@@ -1,11 +1,0 @@
-package com.dataquadinc.model;
-
-public enum UserType {
-
-    ADMIN,
-    SUPERADMIN,
-    EMPLOYEE,
-
-
-
-}
