+++ src/main/java/com/dataquadinc/model/UserType_prod.java
@@ -1,0 +1,11 @@
+package com.dataquadinc.model;
+
+public enum UserType_prod {
+
+    ADMIN,
+    SUPERADMIN,
+    EMPLOYEE,
+
+
+
+}
