+++ src/main/java/com/dataquadinc/repository/LoginRepository.java
@@ -1,10 +1,10 @@
 package com.dataquadinc.repository;
 
-import com.dataquadinc.model.UserDetails;
+import com.dataquadinc.model.UserDetails_prod;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LoginRepository extends JpaRepository<UserDetails, Long> {
-    UserDetails findByEmail(String email);
+public interface LoginRepository extends JpaRepository<UserDetails_prod, Long> {
+    UserDetails_prod findByEmail(String email);
 }
