+++ src/main/java/com/dataquadinc/repository/RolesDao.java
@@ -1,7 +1,7 @@
 package com.dataquadinc.repository;
 
-import com.dataquadinc.model.Roles;
-import com.dataquadinc.model.UserType;
+import com.dataquadinc.model.Roles_prod;
+import com.dataquadinc.model.UserType_prod;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -9,7 +9,7 @@ import java.util.Optional;
 
 
 @Repository
-public interface RolesDao extends JpaRepository<Roles,Integer> {
+public interface RolesDao extends JpaRepository<Roles_prod,Integer> {
 
-   Optional<Roles> findByName(UserType name);
+   Optional<Roles_prod> findByName(UserType_prod name);
 }
