+++ src/main/java/com/dataquadinc/repository/TimesheetRepository.java
@@ -2,17 +2,17 @@ package com.dataquadinc.repository;
 
 
 
-import com.dataquadinc.model.Timesheet;
+import com.dataquadinc.model.Timesheet_prod;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface TimesheetRepository extends JpaRepository<Timesheet, Long> {
+public interface TimesheetRepository extends JpaRepository<Timesheet_prod, Long> {
 
-    Timesheet findByEmployeeIdAndClockOutIsNull(String employeeId);
+    Timesheet_prod findByEmployeeIdAndClockOutIsNull(String employeeId);
 
-    Timesheet findByEmployeeIdAndDate(String employeeId, String date);
-    List<Timesheet> findByEmployeeId(String employeeId);
+    Timesheet_prod findByEmployeeIdAndDate(String employeeId, String date);
+    List<Timesheet_prod> findByEmployeeId(String employeeId);
 
 
 }
