+++ src/main/java/com/dataquadinc/repository/UserDao.java
@@ -1,16 +1,16 @@
 package com.dataquadinc.repository;
 
-import com.dataquadinc.model.UserDetails;
+import com.dataquadinc.model.UserDetails_prod;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import  java.util.*;
 
 @Repository
-public interface UserDao extends JpaRepository<UserDetails,Integer> {
-    UserDetails findByEmail(String email) ;
+public interface UserDao extends JpaRepository<UserDetails_prod,Integer> {
+    UserDetails_prod findByEmail(String email) ;
 
-    UserDetails findByUserId( String userId);
-    UserDetails findByUserName(String userName);
+    UserDetails_prod findByUserId(String userId);
+    UserDetails_prod findByUserName(String userName);
 
-    List<UserDetails> findByRolesId(Long id);
+    List<UserDetails_prod> findByRolesId(Long id);
 }
