+++ src/main/java/com/dataquadinc/repository/leaveCalendarDao.java
@@ -1,9 +1,9 @@
 package com.dataquadinc.repository;
 
-import com.dataquadinc.model.LeaveCalender;
+import com.dataquadinc.model.LeaveCalender_prod;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface leaveCalendarDao extends JpaRepository<LeaveCalender,String> {
+public interface leaveCalendarDao extends JpaRepository<LeaveCalender_prod,String> {
 }
