+++ src/main/java/com/dataquadinc/service/LoginService.java
@@ -3,8 +3,8 @@ package com.dataquadinc.service;
 import com.dataquadinc.dto.LoginDTO;
 import com.dataquadinc.dto.LoginResponseDTO;
 import com.dataquadinc.exceptions.InvalidCredentialsException;
-import com.dataquadinc.model.UserDetails;
-import com.dataquadinc.model.UserType;
+import com.dataquadinc.model.UserDetails_prod;
+import com.dataquadinc.model.UserType_prod;
 import com.dataquadinc.repository.LoginRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
@@ -22,7 +22,7 @@ public class LoginService {
     private BCryptPasswordEncoder passwordEncoder;
 
     public LoginResponseDTO authenticate(LoginDTO loginDTO) {
-        UserDetails userDetails = loginRepository.findByEmail(loginDTO.getEmail());
+        UserDetails_prod userDetails = loginRepository.findByEmail(loginDTO.getEmail());
 
         if (userDetails == null || !passwordEncoder.matches(loginDTO.getPassword(), userDetails.getPassword())) {
             throw new InvalidCredentialsException("Invalid credentials");
@@ -36,7 +36,7 @@ public class LoginService {
 
         userDetails.setLastLoginTime(LocalDateTime.now());
         loginRepository.save(userDetails);
-        UserType roleType = userDetails.getRoles().iterator().next().getName();
+        UserType_prod roleType = userDetails.getRoles().iterator().next().getName();
 
         LoginResponseDTO.Payload payload = new LoginResponseDTO.Payload(
 //                userDetails.getUserId(),
