+++ src/main/java/com/dataquadinc/service/leaveCalenderService.java
@@ -1,7 +1,7 @@
 package com.dataquadinc.service;
 
 import com.dataquadinc.dto.leaveCalenderDto;
-import com.dataquadinc.model.LeaveCalender;
+import com.dataquadinc.model.LeaveCalender_prod;
 import com.dataquadinc.repository.leaveCalendarDao;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -17,8 +17,8 @@ public class leaveCalenderService {
     @Autowired
     private EmailService emailService;  // Inject the EmailService
 
-    public LeaveCalender saveLeave(leaveCalenderDto dto) {
-        LeaveCalender leave = new LeaveCalender();
+    public LeaveCalender_prod saveLeave(leaveCalenderDto dto) {
+        LeaveCalender_prod leave = new LeaveCalender_prod();
         leave.setUserId(dto.getUserId());
         leave.setManagerEmail(dto.getManagerEmail());
         leave.setDescription(dto.getDescription());
@@ -29,7 +29,7 @@ public class leaveCalenderService {
 
 
         // Save the leave
-        LeaveCalender savedLeave = leaveCalendarDaoo.save(leave);
+        LeaveCalender_prod savedLeave = leaveCalendarDaoo.save(leave);
 
 // Send email to manager after leave is saved
         String managerEmail = dto.getManagerEmail(); // Assuming you have the manager's email in the DTO
@@ -49,7 +49,7 @@ public class leaveCalenderService {
 
         return savedLeave;
     }
-    public List<LeaveCalender> getAllLeaves() {
+    public List<LeaveCalender_prod> getAllLeaves() {
         return leaveCalendarDaoo.findAll(); // Fetch all leave records
     }
 }
