+++ src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.datasource.url=jdbc:mysql://104.197.113.135:3306/dataquad
+spring.datasource.url=jdbc:mysql://104.197.113.135:3306/dataquad-prod
 spring.datasource.username=root
 spring.datasource.password=Welcome@1234
 spring.jpa.hibernate.ddl-auto=update
